Remove uint32_t stride casts after nozzle C API promotion to int64_t
nozzle_swizzle_channels, nozzle_widen_uint16_to_uint32,
nozzle_widen_half_to_float now accept int64_t stride params.
Remove static_cast<uint32_t> from send and receive callers.

Nozzle submodule updated to a80de6b.

+++ source/projects/jit.nozzle.send/jit.nozzle.send.cpp
@@ -240,7 +240,7 @@ private:
 
 				NozzleErrorCode swiz_err = nozzle_swizzle_channels(
 					src, dst, w, h,
-					matrix_row_bytes, static_cast<uint32_t>(mapped.row_stride_bytes),
+					matrix_row_bytes, mapped.row_stride_bytes,
 					swiz_fmt, permute_map);
 				if (swiz_err != NOZZLE_OK) {
 					cerr << "jit.nozzle.send: swizzle failed (error " << swiz_err << ")" << endl;
